prepare for MM_MODULES_DIR and MM_CUSTOMCSS_FILE

+++ build/entrypoint.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
 base="/opt/magic_mirror"
+mounted="$(mount | sed -rn 's|.*on\s*'$base'([^ ]*).*|'$base'\1|p' | xargs)"
+
 modules_dir="${base}/modules"
 default_dir="${modules_dir}/default"
 config_dir="${base}/config"
@@ -14,11 +16,15 @@ _error() {
   echo "[entrypoint $(date +%T.%3N)] [ERROR]  $1"
 }
 
+_is_mounted() {
+  echo "$mounted" | grep -Eq ${base}'/'${1}
+}
+
 _start_mm() {
   if [ "$(id -u)" = "0" ]; then
     _info "running as root but starting the magicmirror process with uid=1000"
     # directories must be writable by user node:
-    chown -R node:node ${modules_dir} ${config_dir} ${css_dir}
+    chown -R node:node ${mounted}
     _file="mm.env"
     rm -f $_file
     echo "export START_CMD=\"$@\"" > $_file
@@ -33,15 +39,10 @@ _start_mm() {
   fi
 }
 
-# directories should be mounted, if not, create them:
-mkdir -p ${modules_dir}
-mkdir -p ${config_dir}
-mkdir -p ${css_dir}
-
 if [ "$STARTENV" = "init" ]; then
-  _info "change permissions for folders modules, config, css ..."
-  chown -R ${MM_UID}:${MM_GID} ${modules_dir} ${config_dir} ${css_dir}
-  chmod -R ${MM_CHMOD} ${modules_dir} ${config_dir} ${css_dir}
+  _info "change permissions for folders ${mounted} ..."
+  chown -R ${MM_UID}:${MM_GID} ${mounted}
+  chmod -R ${MM_CHMOD} ${mounted}
   _info "done."
 
   exit 0
@@ -65,36 +66,47 @@ if [ -z "$TZ" ]; then
   _info "***WARNING*** could not set timezone, please set TZ variable in compose.yaml, see https://khassel.gitlab.io/magicmirror/configuration/#timezone"
 fi
 
-if [ ! -d "${default_dir}" ]; then
-  MM_OVERRIDE_DEFAULT_MODULES=true
-  mkdir -p ${default_dir}
-fi
-
-if [ "${MM_OVERRIDE_DEFAULT_MODULES}" = "true" ]; then
-  if [ -w "${default_dir}" ]; then
-    _info "copy default modules"
-    rm -rf ${default_dir}
+if _is_mounted "modules"; then
+  if [ ! -d "${default_dir}" ]; then
+    MM_OVERRIDE_DEFAULT_MODULES=true
     mkdir -p ${default_dir}
-    cp -r ${base}/mount_ori/modules/default/. ${default_dir}/
-  else
-    _error "No write permission for ${default_dir}, skipping copying default modules"
   fi
+
+  if [ "${MM_OVERRIDE_DEFAULT_MODULES}" = "true" ]; then
+    if [ -w "${default_dir}" ]; then
+      _info "copy default modules"
+      rm -rf ${default_dir}
+      mkdir -p ${default_dir}
+      cp -r ${base}/mount_ori/modules/default/. ${default_dir}/
+    else
+      _error "No write permission for ${default_dir}, skipping copying default modules"
+    fi
+  fi
+else
+  mkdir -p ${modules_dir}
+  [ -d "${base}/mount_ori/modules/default" ] && mv ${base}/mount_ori/modules/default ${modules_dir}
 fi
 
-[ ! -f "${css_dir}/main.css" ] && MM_OVERRIDE_CSS=true
+if _is_mounted "css"; then
+  [ ! -f "${css_dir}/main.css" ] && MM_OVERRIDE_CSS=true
 
-if [ "${MM_OVERRIDE_CSS}" = "true" ]; then
-  if [ -w "${css_dir}" ]; then
-    _info "copy css files"
-    cp ${base}/mount_ori/css/* ${css_dir}/
-    # create css/custom.css file https://github.com/MagicMirrorOrg/MagicMirror/issues/1977
-    [ ! -f "${css_dir}/custom.css" ] && touch ${css_dir}/custom.css
-  else
-    _error "No write permission for ${css_dir}, skipping copying css files"
+  if [ "${MM_OVERRIDE_CSS}" = "true" ]; then
+    if [ -w "${css_dir}" ]; then
+      _info "copy css files"
+      cp ${base}/mount_ori/css/* ${css_dir}/
+      # create css/custom.css file https://github.com/MagicMirrorOrg/MagicMirror/issues/1977
+      [ ! -f "${css_dir}/custom.css" ] && touch ${css_dir}/custom.css
+    else
+      _error "No write permission for ${css_dir}, skipping copying css files"
+    fi
   fi
+else
+  [ -d "${base}/mount_ori/css" ] && mv ${base}/mount_ori/css ${base}
+  touch ${css_dir}/custom.css
 fi
 
 if [ ! -f "${config_dir}/config.js" ]; then
+  mkdir -p ${config_dir}
   if [ -w "${config_dir}" ]; then
     _info "copy default config.js"
     cp ${base}/mount_ori/config/config.js.sample ${config_dir}/config.js
